+++ Marlin/cap_sense_probe.h
@@ -39,7 +39,7 @@ struct ProbeResult
  *  Probe the bed with the capacitive sensor. Report back the Z position where the bed is hit.
  *  Does not change the X/Y position. Sets the Z position on the reported Z position.
  */
-ProbeResult probeWithCapacitiveSensor(const float start_position[], const int feedrate, const int verbosity, const float move_distance, const float extra_z_move);
+ProbeResult probeWithCapacitiveSensor(const float start_position[], const int16_t feedrate, const int16_t verbosity, const float move_distance, const float extra_z_move);
 
 /**
  *  Make a move in the Z direction only while logging alle capacitive sensor data.
@@ -47,17 +47,17 @@ ProbeResult probeWithCapacitiveSensor(const float start_position[], const int fe
  *  @param feedrate: speed of which to move in mm/minute
  *  @param move_distance: distance moved in Z that the bed is moved upwards, in mm.
  */
-void moveWithCapacitiveSensor(const int feedrate, const float move_distance);
+void moveWithCapacitiveSensor(const int16_t feedrate, const float move_distance);
 
 /**
  * Compensate any static capacitance by setting the CAPDAC. Finds best value by increasing until
  * measured capacitance is closest to zero, but still positive.
  */
-void calibrateCapacitanceOffset(int verbosity);
+void calibrateCapacitanceOffset(int16_t verbosity);
 
 /**
  * sets the base capacitive offsets based on the average of a couple of current samples from the capacitive sensor
  */
-void updateCapacitiveSensorBaseline(int verbosity);
+void updateCapacitiveSensorBaseline(int16_t verbosity);
 
 #endif//CAP_SENSE_PROBE_H
